ProfileSettings.tsx changed to AccountSettings.tsx
